+++ packages/sanity/src/core/components/collapseMenu/__workshop__/CollapseMenuStory.tsx
@@ -2,7 +2,7 @@ import {IceCreamIcon} from '@sanity/icons'
 import {Card, Flex} from '@sanity/ui'
 import {useBoolean, useSelect} from '@sanity/ui-workshop'
 import React from 'react'
-import {ContextMenuButton} from '../../../../ui/contextMenuButton'
+import {ContextMenuButton} from '../../contextMenuButton'
 import {CollapseMenu, CollapseMenuButton} from '../'
 
 const GAP_OPTIONS = {'0': 0, '1': 1, '2': 2, '3': 3, '4': 4}