+++ packages/sanity/src/core/form/inputs/PortableText/toolbar/ActionMenu.tsx
@@ -8,7 +8,7 @@ import {
 import {isKeySegment} from '@sanity/types'
 import {PopoverProps} from '../../../../../ui'
 import {CollapseMenu, CollapseMenuButton} from '../../../../components/collapseMenu'
-import {ContextMenuButton} from '../../../../../ui/contextMenuButton'
+import {ContextMenuButton} from '../../../../components/contextMenuButton'
 import {PTEToolbarAction, PTEToolbarActionGroup} from './types'
 import {useActiveActionKeys, useFocusBlock} from './hooks'
 import {getActionIcon} from './helpers'