+++ packages/sanity/src/core/form/studio/assetSource/shared/AssetMenu.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import {MenuButton, MenuItem} from '../../../../../ui'
 import {useTranslation} from '../../../../i18n'
 import {AssetMenuAction} from '../types'
-import {ContextMenuButton} from '../../../../../ui/contextMenuButton'
+import {ContextMenuButton} from '../../../../components/contextMenuButton'
 
 const MENU_POPOVER_PROPS = {portal: true, placement: 'right'} as const