+++ packages/sanity/src/desk/components/pane/__workshop__/SplitPanesStory/DocumentPane/DocumentActions.tsx
@@ -2,7 +2,7 @@ import {PublishIcon} from '@sanity/icons'
 import {Flex, Menu, Stack, useToast} from '@sanity/ui'
 import React, {useCallback} from 'react'
 import {Button, MenuButton, MenuButtonProps, MenuItem} from '../../../../../../ui'
-import {ContextMenuButton} from '../../../../../../ui/contextMenuButton'
+import {ContextMenuButton} from 'sanity'
 
 const MENU_BUTTON_POPOVER_PROPS: MenuButtonProps['popover'] = {
   constrainSize: true,